requirements.txt
streamlit
langchain-core
langchain-google-genai
datetime